[wip] added intermediate ChunckReader
